+++ 20170919/颜番-151220139/FancyYan/HuluSort/HuluBaby.java
@@ -1,3 +1,11 @@
+package FancyYan.HuluSort;
+
+/**
+ * 此文件包含葫芦娃的类HuluBaby
+ * 葫芦娃拥有排行这一个成员变量，可以报告自己的排行，颜色（字符串）
+ * 当进行位置交换时，葫芦娃可以报告交换位置的信息（直接输出）
+ */
+
 public class HuluBaby {
     int rank;
 
@@ -5,10 +13,18 @@ public class HuluBaby {
         rank = _rank;
     }
 
+    /**
+     * 外部获取该葫芦娃的排行
+     * @return 葫芦娃的排行值(int)
+     */
     public int getRank(){
         return rank;
     }
 
+    /**
+     * 以字符形式报告葫芦娃的排行，没有输出
+     * @return 葫芦娃的排行说明，如"大娃：1"形式的字符串
+     */
     public String reportRank(){
         String output="";
         switch(rank){
@@ -23,6 +39,10 @@ public class HuluBaby {
         return output;
     }
 
+    /**
+     * 以字符串形式报告葫芦娃的颜色特征
+     * @return 葫芦娃颜色字符串，如”红色“
+     */
     public String reportColor(){
         String output="";
         switch(rank){
@@ -36,9 +56,14 @@ public class HuluBaby {
         }
         return output;
     }
+
+    /**
+     * 报告排序时的位置交换，以屏幕输出的方式
+     * @param oldPos 原来的位置
+     * @param newPos 交换之后的位置
+     */
     public void reportExchangePos(int oldPos, int newPos){
         System.out.printf("%s: %d --> %d\n", reportRank(),oldPos, newPos);
     }
 
 }
-
